ref: Rename stimuli.js as baseStimulus.js
