+++ src/JsPsych/trials/examples/camera/CameraStart.js
@@ -2,7 +2,7 @@ import htmlButtonResponse from '@jspsych/plugin-html-button-response';
 
 // TODO: Move markup to JsPsych?
 import { photodiodeGhostBox } from '../../../markup/photodiode';
-import { baseStimulus } from '../../../markup/stimuli';
+import { baseStimulus } from '../../../markup/baseStimulus';
 
 import { language } from '../../../language';
 import { TASK_NAME } from '../../../constants';
