+++ src/JsPsych/trials/examples/partials/baseStimulus.js
