+++ src/index.css
@@ -1,21 +1,24 @@
 /* GLOBALS */
 
 body {
-  font-family: -apple-system, BlinkMacSystemFont, "Segoe UI", "Roboto", "Oxygen",
-    "Ubuntu", "Cantarell", "Fira Sans", "Droid Sans", "Helvetica Neue",
-    sans-serif;
+  font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', 'Roboto', 'Oxygen', 'Ubuntu',
+    'Cantarell', 'Fira Sans', 'Droid Sans', 'Helvetica Neue', sans-serif;
   -webkit-font-smoothing: antialiased;
   -moz-osx-font-smoothing: grayscale;
 }
 
 code {
-  font-family: source-code-pro, Menlo, Monaco, Consolas, "Courier New", monospace;
+  font-family: source-code-pro, Menlo, Monaco, Consolas, 'Courier New', monospace;
 }
 
-p, li {
+p,
+li {
   font-size: 1.5rem;
 }
-html, body, #root, .Experiment {
+html,
+body,
+#root,
+.Experiment {
   text-align: center;
   margin: 0;
   min-height: 100vh;
@@ -30,7 +33,7 @@ html, body, #root, .Experiment {
   flex-direction: column;
   justify-content: center;
   align-items: center;
-  width:100%;
+  width: 100%;
   height: 100%;
 }
 
@@ -42,7 +45,8 @@ html, body, #root, .Experiment {
   width: 100%;
 }
 
-.main-prompt, .center_container {
+.main-prompt,
+.center_container {
   height: 100vh;
   display: flex;
   flex-direction: row;
@@ -72,7 +76,7 @@ html, body, #root, .Experiment {
   bottom: 3px;
   right: 3px;
   display: flex;
-  outline: 3px solid rgb(80,80,80);
+  outline: 3px solid rgb(80, 80, 80);
   margin: auto;
   height: 50px;
   width: 50px;
@@ -86,6 +90,7 @@ html, body, #root, .Experiment {
   position: absolute;
 }
 
+/* TODO: Not used? fixation() from behavior-task-trials needs it*/
 #fixation-dot {
   background-color: black;
   height: 15px;
